Improve readability of `sleepTimer` creation
Before, the significance of the `time.Second` argument to
`time.NewTimer` was not clear, and I thought it was odd that
`sleepTimerActive` was initialized to `false` even though the timer
was initially active.

At first I thought there might be a theoretical (but highly unlikely)
race condition: What would happen if the timer expired (and the
current time sent to `sleepTimer.C`) between the `time.NewTimer()`
call and the `sleepTimer.Stop()` call?  After careful study I
concluded that it would still behave correctly, but it certainly
wasn't obvious to me.  Now I hope that it is obvious to all readers
that the initialization behavior is correct.